Random Password Generator.ipynb
This code generates a random password of length 8 (which can be changed by specifying a different length when calling the generate_password function), and then prints the password to the console. The user is prompted to enter whether they want to generate another password or not. If they enter "y", the loop continues and generates another password; otherwise, the loop ends and the program terminates.